Finished 2nd exercise
Script stops running when pattern is not matched in the second parameter.
Also, the .tmp file is deleted when an EXIT, SIGQUIT, SIGKILL or SIGTERM signal is sent to the script.

Also, the awk command is the one to go for getting the 5th and the 9th column for the $size and $path parameters to correctly and consistently use.